+++ CHANGELOG.md
@@ -6,6 +6,7 @@ changes to any options/interfaces the checker exposes for developers.
 NN XXX 2018
   - Ensure vnu.jar is always runnable under Java8, even if built under Java9
   - Align ARIA checking further with ARIA in HTML spec requirements
+  - Restore the language-detection feature to vnu.jar command-line checker
 
 # 18.3.0
 24 March 2018